Delete the orphaned /testsetups/new upload path (#1119)
WebRoutes.createSetup duplicated TestSetupRoutes.uploadTestSetup but had
drifted behind it: no zip-bomb guard (validateZipUploadSize), no
dependency-graph validation, no grading-mode checks, no browser-mode
flat-notebook extraction. No template has linked to /testsetups/new since
the draft-based new-assignment flow shipped — it was reachable only by
direct URL, which is exactly how a hardening fix lands on one intake path
and misses the twin.

- GET/POST /testsetups/new, both handlers, and setup-new.leaf are deleted;
  programmatic uploads go through the hardened POST /api/v1/testsetups.
- The auth test probing the old page now probes the /instructor area gate.
- While here (from the issue): the pure notebook-JSON and zip-inspection
  helpers (filterNotebook, mergeNotebook, normalizeNotebookForJupyterLite,
  notebookData(for:), zipContainsNotebook, extractNotebookFromZip, …) move
  out of TestSetupRoutes.swift to Helpers/NotebookContentHelpers.swift —
  they are content utilities used across routes/services/MCP, not routing.
  TestSetupRoutes.swift drops 607 → ~356 lines. No behavior change.
